+++ js/src/widget.js
@@ -49,7 +49,12 @@ export class AutocompleteWidget {
 
     static initialize(context) {
         $('div.yafowil-widget-autocomplete', context).each(function() {
-            new AutocompleteWidget($(this));
+            let elem = $(this);
+            if (window.yafowil_array !== undefined &&
+                window.yafowil_array.inside_template(elem)) {
+                return;
+            }
+            new AutocompleteWidget(elem);
         });
     }
 
@@ -337,5 +342,23 @@ export class AutocompleteWidget {
 
     hide_dropdown() {
         this.dd_elem.hide();
+    }
+}
+
+//////////////////////////////////////////////////////////////////////////////
+// yafowil.widget.array integration
+//////////////////////////////////////////////////////////////////////////////
+
+function autocomplete_on_array_add(inst, context) {
+    AutocompleteWidget.initialize(context);
+}
+
+export function register_array_subscribers() {
+    if (window.yafowil_array !== undefined) {
+        window.yafowil_array.on_array_event('on_add', autocomplete_on_array_add);
+    } else if (yafowil.array !== undefined) {
+        $.extend(yafowil.array.hooks.add, {
+            autocomplete_binder: AutocompleteWidget.initialize
+        });
     }
 }