+++ js/tests/test_widget.js
@@ -1,4 +1,5 @@
 import {AutocompleteSuggestion, AutocompleteWidget} from '../src/widget.js';
+import {register_array_subscribers} from '../src/widget.js';
 
 let container = $('<div id="container" />');
 
@@ -7,6 +8,9 @@ QUnit.module('AutocompleteWidget', hooks => {
     let widget;
     let arr = "|one|two|three|four|";
     let params = "delay,0|minLength,0|type,local";
+    let _array_subscribers = {
+        on_add: []
+    };
 
     hooks.beforeEach(() => {
         $('body').append(container);
@@ -33,6 +37,52 @@ QUnit.module('AutocompleteWidget', hooks => {
         assert.ok(widget.hide_dropdown);
         assert.ok(widget.on_keydown);
         assert.ok(widget.autocomplete);
+    });
+
+    QUnit.test('register_array_subscribers', assert => {
+        container.empty();
+        // patch window.yafowil
+        window.yafowil = {
+            array: undefined
+        }
+        // return if window.yafowil === undefined
+        register_array_subscribers();
+        widget = elem.data('yafowil-autocomplete');
+
+        // patch yafowil_array
+        window.yafowil_array = {
+            on_array_event: function(evt_name, evt_function) {
+                _array_subscribers[evt_name] = evt_function;
+            },
+            inside_template(elem) {
+                return elem.parents('.arraytemplate').length > 0;
+            }
+        }
+        register_array_subscribers();
+
+        // create table DOM
+        let table = $('<table />')
+            .append($('<tr id="row" />'))
+            .append($('<td />'))
+            .appendTo('body');
+        $('td', table).addClass('arraytemplate');
+        let el = create_elem(arr, params);
+        el.appendTo($('td', table));
+
+        // invoke array on_add - returns
+        let context = $('#row');
+        _array_subscribers['on_add'].apply(null, context);
+        widget = el.data('yafowil-autocomplete');
+        assert.notOk(widget);
+        $('td', table).removeClass('arraytemplate');
+
+        // invoke array on_add
+        _array_subscribers['on_add'].apply(null, context);
+        widget = el.data('yafowil-autocomplete');
+        assert.ok(widget);
+
+        table.remove();
+        delete window.yafowil;
     });
 
     QUnit.test('parse_options', assert => {
@@ -463,7 +513,10 @@ QUnit.module('AutocompleteWidget', hooks => {
                 assert.strictEqual(widget.current_focus, 0);
                 // trigger Enter key
                 widget.input_elem.trigger(enter);
-                assert.strictEqual(widget.input_elem.val(), widget.suggestions[0].value);
+                assert.strictEqual(
+                    widget.input_elem.val(),
+                    widget.suggestions[0].value
+                );
                 done();
             }, 10);
         });