+++ src/yafowil/widget/autocomplete/resources/widget.js
@@ -42,7 +42,12 @@ var yafowil_autocomplete = (function (exports, $) {
     class AutocompleteWidget {
         static initialize(context) {
             $('div.yafowil-widget-autocomplete', context).each(function() {
-                new AutocompleteWidget($(this));
+                let elem = $(this);
+                if (window.yafowil_array !== undefined &&
+                    window.yafowil_array.inside_template(elem)) {
+                    return;
+                }
+                new AutocompleteWidget(elem);
             });
         }
         constructor(elem) {
@@ -295,6 +300,18 @@ var yafowil_autocomplete = (function (exports, $) {
             this.dd_elem.hide();
         }
     }
+    function autocomplete_on_array_add(inst, context) {
+        AutocompleteWidget.initialize(context);
+    }
+    function register_array_subscribers() {
+        if (window.yafowil_array !== undefined) {
+            window.yafowil_array.on_array_event('on_add', autocomplete_on_array_add);
+        } else if (yafowil.array !== undefined) {
+            $.extend(yafowil.array.hooks.add, {
+                autocomplete_binder: AutocompleteWidget.initialize
+            });
+        }
+    }
 
     $(function() {
         if (window.ts !== undefined) {
@@ -304,15 +321,12 @@ var yafowil_autocomplete = (function (exports, $) {
         } else {
             AutocompleteWidget.initialize();
         }
-        if (yafowil.array !== undefined) {
-            $.extend(yafowil.array.hooks.add, {
-                autocomplete_binder: AutocompleteWidget.initialize
-            });
-        }
+        register_array_subscribers();
     });
 
     exports.AutocompleteSuggestion = AutocompleteSuggestion;
     exports.AutocompleteWidget = AutocompleteWidget;
+    exports.register_array_subscribers = register_array_subscribers;
 
     Object.defineProperty(exports, '__esModule', { value: true });