+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changes
 2.0 (unreleased)
 ----------------
 
+- Extend JS by ``autocomplete_on_array_add`` and ``register_array_subscribers``
+  to enable usage in ``yafowil.widget.array``.
+  [lenadax]
+
 - Get rid of jQuery UI. Rewrite Jacascript widget using ES6.
   [lenadax]